+++ test/test_helper.exs
@@ -1,3 +1,1 @@
 ExUnit.start()
-
-Mox.defmock(MockSender, for: Tapper.Reporter.Api)